BE/requirements.txt
alembic==1.11.1
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
distlib==0.3.6
filelock==3.12.2
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
PyJWT==2.6.0
Mako==1.2.4
MarkupSafe==2.1.3
platformdirs==2.6.2
psycopg2-binary==2.9.6
pycparser==2.21
six==1.16.0
SQLAlchemy==1.4.41
typing_extensions==4.7.1
virtualenv==20.15.1
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.15.0
